+++ Calc2/CalculatorPresenter.java
@@ -28,39 +28,39 @@ public void choice() {
     }           
 }
 public void onAddButtonClicked() {
-    int number1 = view.getUserInput();
-    int number2 = view.getUserInput();
+    double number1 = view.getUserInput();
+    double number2 = view.getUserInput();
 
     model.add(number1, number2);
-    int result = model.getResult();
+    double result = model.getResult();
 
     view.displayResult(result);
 }
 
 public void onSubtracButtonClicked() {
-    int number1 = view.getUserInput();
-    int number2 = view.getUserInput();
+    double number1 = view.getUserInput();
+    double number2 = view.getUserInput();
 
     model.subtrac(number1, number2);
-    int result = model.getResult();
+    double result = model.getResult();
 
     view.displayResult(result);
 }
 public void onMultipButtonClicked() {
-    int number1 = view.getUserInput();
-    int number2 = view.getUserInput();
+    double number1 = view.getUserInput();
+    double number2 = view.getUserInput();
 
     model.multip(number1, number2);
-    int result = model.getResult();
+    double result = model.getResult();
 
     view.displayResult(result);
 }
 public void onDivisButtonClicked() {
-    int number1 = view.getUserInput();
-    int number2 = view.getUserInput();
+    double number1 = view.getUserInput();
+    double number2 = view.getUserInput();
 
     model.divis(number1, number2);
-    int result = model.getResult();
+    double result = model.getResult();
 
     view.displayResult(result);
 }
